+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+#### 2.0.2
+Add function to format number to international format *2017-09-22*
+Allow for `postData` to be overridden from the calling function
+
 #### 2.0.1
 * Rename flash messages JavaScript variable in BaseController *2017-09-18*